pyproject.toml
[tool.poetry]
name = "batamp-viz"
version = "0.2.0"
description = "Bat Acoustic Monitoring Visualization Tool"
authors = ["Brendan Ward <bcward@astutespruce.com>"]
license = "MIT"
packages = [{ include = "analysis" }]

[tool.poetry.dependencies]
python = "^3.10"
pyogrio = "^0.9.0"
python-dotenv = "^1.0.1"
pyarrow = "^16.1.0"
python-databasin = "^0.6.1"
pymgl = "^0.4.0"
shapely = "^2.0.2"
ruff = "^0.4.9"
geopandas = "^0.14.4"
httpx = {extras = ["http2"], version = "^0.27.0"}
numba = "^0.60.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
